refactor!: rename Int32 to Int and Int64 to BigInt
Embrace JavaScript semantics for integer types:
- Int32 -> Int (maps to JS number, i64 storage in Rust)
- Int64 -> BigInt (maps to JS BigInt)

Key changes:
- Remove artificial Int32 overflow checking in codegen
- Int division now uses $div helper for truncation and div-by-zero
- Int accepts full i64 range (no longer limited to i32)
- Update all type annotations, methods, and 523 tests

BREAKING CHANGE: Type names changed from Int32/Int64 to Int/BigInt

+++ src/ir.rs
@@ -141,8 +141,8 @@ pub struct TypedMatchArm {
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum TypedExpr {
-    Int32(i32),
-    Int64(i64),
+    Int(i64),
+    BigInt(i64),
     Float(f64),
     Bool(bool),
     String(String),
@@ -229,8 +229,8 @@ pub enum TypedEnumConstructFields {
 impl TypedExpr {
     pub fn ty(&self) -> Type {
         match self {
-            TypedExpr::Int32(_) => Type::Int32,
-            TypedExpr::Int64(_) => Type::Int64,
+            TypedExpr::Int(_) => Type::Int,
+            TypedExpr::BigInt(_) => Type::BigInt,
             TypedExpr::Float(_) => Type::Float,
             TypedExpr::Bool(_) => Type::Bool,
             TypedExpr::String(_) => Type::String,
